bghyn vtjju78yjhiujhyn7i8m

+++ PGU2023/Assets/Scripts/EndWriter.cs
@@ -8,7 +8,7 @@ public class EndWriter : MonoBehaviour
     [SerializeField] EndMoveData data;
     [SerializeField] MoveData startdata;
 
-    public void StoreData(int cost, int value, int intrest, int functionality)
+    public void StoreData(int cost, int value, float intrest, int functionality)
     {
         data.cost = cost;
         data.value = value;
